refactor(frontend): разнос ApiClient (2781 строка, ~300 методов) на доменные слои
Этап F2: lib/api/client.ts разнесён по #region-границам на цепочку
классов-слоёв (методы перенесены дословно, private → protected):

CoreClient (clientCore.ts — транспорт request, 401/сеть/gateway,
mock-хелперы) → TunnelsClient (туннели, control, импорт, ping check,
внешние и системные туннели) → SystemClient (testing, system, updates,
settings, auth, boot, logging, terminal, signature, connections,
dns check, diagnostics, monitoring) → ServersClient (VPN/managed
серверы, backup) → RoutingClient (static/dns routes, политики,
client routes) → SingboxClient (sing-box, device proxy) →
SbRouterClient (router, selective, fakeip) → SubscriptionsClient.

client.ts стал фасадом: ApiClient extends SubscriptionsClient,
export const api + реэкспорт ApiGatewayError/TrafficPeriod — все 133
импортёра $lib/api/client не тронуты.

Эквивалентность верифицирована мультимножествами строк кода и
комментариев: расхождения — только формат списков импортов, намеренные
export у ApiResponse/readDiagnosticsSanitizedPreference и заголовок
класса. svelte-check 0/0, eslint 0, vitest 1343, build зелёный.
